+++ src/main/java/org/sopt/exception/TitleLengthException.java
@@ -1,7 +1,7 @@
 package org.sopt.exception;
 
-import org.sopt.global.exception.BusinessException;
-import org.sopt.global.exception.ExceptionCode;
+import org.sopt.global.config.exception.BusinessException;
+import org.sopt.global.config.exception.constant.ExceptionCode;
 
 public class TitleLengthException extends BusinessException {
 	public TitleLengthException() {
