+++ src/main/java/org/sopt/global/config/exception/BusinessException.java
@@ -1,4 +1,6 @@
-package org.sopt.global.exception;
+package org.sopt.global.config.exception;
+
+import org.sopt.global.config.exception.constant.ExceptionCode;
 
 public class BusinessException extends RuntimeException {
